refactoring

+++ src/Evaluator.cpp
@@ -829,7 +829,7 @@ namespace Evaluator {
                 }
 
                 // If non-pawn pieces differ
-                eg_value *= 1.0 + (double)(nonpawn_count[Own]-nonpawn_count[Opp]) / (nonpawn_count[Own]+nonpawn_count[Opp]+2) / 2;
+                eg_value *= 1.0 + (nonpawn_count[Own]-nonpawn_count[Opp]) / (nonpawn_count[Own]+nonpawn_count[Opp]+2) / 2.0;
 
                 score += mk_score (mg_value, eg_value) + PAWN_PASSED_FILE[_file (s)];
             }
